+++ lib/generators/modulorails/docker/docker_generator.rb
@@ -17,11 +17,10 @@ class Modulorails::DockerGenerator < Modulorails::Generators::Base
     @image_name = @data.name.parameterize
     @environment_name = @data.environment_name
 
-    template 'Dockerfile'
-    template 'Dockerfile.prod'
-    template 'compose.yml'
-    template 'entrypoints/docker-entrypoint.sh', 'bin/docker-entrypoint'
-    chmod 'bin/docker-entrypoint', 0o755
+    create_dockerfile
+    create_dockerfile_prod
+    create_compose_yml
+    create_docker_entrypoint
     template 'config/database.yml'
     template 'config/cable.yml'
     template 'config/initializers/0_redis.rb'
@@ -30,10 +29,50 @@ class Modulorails::DockerGenerator < Modulorails::Generators::Base
     # Useless unless project is using Webpacker
     return unless @webpack_container_needed
 
-    template 'entrypoints/webpack-entrypoint.sh', 'bin/webpack-entrypoint'
-    chmod 'bin/webpack-entrypoint', 0o755
+    create_webpack_entrypoint
   rescue StandardError => e
     warn("[Modulorails] Error: cannot generate Docker configuration: #{e.message}")
+  end
+
+  private
+
+  def create_dockerfile
+    say('WARNING: The entrypoint was moved. Check that your Dockerfile still works.') if File.exist?('Dockerfile')
+    template 'dockerfiles/modulotech/Dockerfile', 'Dockerfile'
+  end
+
+  def create_webpack_entrypoint
+    create_new_file('entrypoints/webpack-entrypoint.sh', 'bin/webpack-entrypoint')
+  end
+
+  def create_docker_entrypoint
+    create_new_file 'entrypoints/docker-entrypoint.sh', 'bin/docker-entrypoint'
+  end
+
+  def create_compose_yml
+    create_new_file 'docker-compose.yml', 'compose.yml', executable: false
+  end
+
+  def create_dockerfile_prod
+    if File.exist?('Dockerfile.prod')
+      say('WARNING: The entrypoint was moved. Check that your Dockerfile.prod still works.')
+    end
+
+    if Gem::Version.new(@data.rails_version) >= Gem::Version.new('7.2')
+      template 'dockerfiles/rails/Dockerfile.prod', 'Dockerfile.prod'
+    else
+      template 'dockerfiles/modulotech/Dockerfile.prod', 'Dockerfile.prod'
+    end
+  end
+
+  def create_new_file(old_file, new_file, executable: true)
+    if File.exist?(old_file)
+      copy_file old_file, new_file
+      remove_file old_file
+    else
+      template old_file, new_file
+    end
+    chmod new_file, 0o755 if executable
   end
 
 end